requirements.txt
Flask
scikit-learn
pandas
numpy
Pillow
gunicorn
